+++ src/interfaces/markdown-page-collection-interface.ts
@@ -2,4 +2,5 @@ import { IMarkdownPage } from "./markdown-page-interface";
 
 export interface IMarkdownPageCollection extends IMarkdownPage {
 	children: IMarkdownPage[];
+	log(): void;
 }