+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- `mdPagesLabel` option for specifying the top-level page title
+- Improved navigation handling and breadcrumbs
+- Better logging for discovered pages
 
 ## [0.1.0] - 2019-04-14
 ### Added
 - Initial commit of the plugin with README
 - This CHANGELOG file.
 
+[Unreleased]: https://github.com/mipatterson/typedoc-plugin-markdown-pages/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/mipatterson/typedoc-plugin-markdown-pages/releases/tag/v0.1.0